pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "motd-analyser"
version = "0.1.0"
description = "Automated video analysis pipeline to measure coverage bias in BBC's Match of the Day"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Michael"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "opencv-python==4.12.0.88",
    "numpy>=2.0.0,<2.4.0",
    "pyyaml==6.0.3",
    "scenedetect[opencv]==0.6.7.1",
    "easyocr==1.7.2",
    "torch==2.9.0",
    "torchvision==0.24.0",
    "torchaudio==2.9.0",
    "faster-whisper==1.1.0",
    "requests>=2.32.0",
    "click==8.1.8",
    "python-dateutil==2.9.0.post0",
    "tqdm==4.67.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.2",
    "pytest-cov==7.0.0",
]

[project.scripts]
motd = "motd.__main__:cli"

[tool.setuptools.packages.find]
where = ["src"]
